requirements.txt
cython
numpy
opencv-python
opencv-contrib-python
pyopengl
scipy
pyntcore

ultralytics